Hive pattern
