+++ tests/patterns/test_hive.py
@@ -142,12 +142,12 @@ def test_hive_sel_core_1():
     root = earthkit_test_data_file("pattern/1")
     pattern = "{variable}_{base_date:date(%Y-%m-%dT-H-%M)}_{step}.grib"
 
-    ds = from_source("file-pattern", os.path.join(root, pattern), hive_partitioning=True).to_source()
+    ds = from_source("file-pattern", os.path.join(root, pattern), hive_partitioning=True)
 
     # assert ds.root == path
     diag = HiveDiag()
     # using hive partitioning keys
-    r = ds.sel(variable="t", step=12, _hive_diag=diag)
+    r = ds.to_fieldlist(variable="t", step=12, _hive_diag=diag)
     assert diag.file_count == 1
     assert diag.sel_count == 0
     assert len(r) == 6
@@ -166,7 +166,7 @@ def test_hive_sel_core_1():
 
     # using hive partitioning keys + extra keys from GRIB header
     diag.reset()
-    r = ds.sel(
+    r = ds.to_fieldlist(
         {
             "variable": "t",
             "step": 12,
@@ -183,9 +183,9 @@ def test_hive_sel_core_2():
     root = earthkit_test_data_file("pattern/invalid")
     pattern = "_{shortName}_{date:date(%Y-%m-%dT%H:%M)}_{step}.grib"
 
-    ds = from_source("file-pattern", os.path.join(root, pattern), hive_partitioning=True).to_source()
+    ds = from_source("file-pattern", os.path.join(root, pattern), hive_partitioning=True)
 
-    r = ds.sel(variable="t", step=12)
+    r = ds.to_fieldlist(variable="t", step=12)
     assert len(r) == 0
 
 
@@ -193,12 +193,12 @@ def test_hive_sel_grib_1():
     root = earthkit_test_data_file("pattern/1")
     pattern = "{shortName}_{date:date(%Y-%m-%dT-H-%M)}_{step}.grib"
 
-    ds = from_source("file-pattern", os.path.join(root, pattern), hive_partitioning=True).to_source()
+    ds = from_source("file-pattern", os.path.join(root, pattern), hive_partitioning=True)
 
     # assert ds.root == path
     diag = HiveDiag()
     # using hive partitioning keys
-    r = ds.sel(shortName="t", step=12, _hive_diag=diag)
+    r = ds.to_fieldlist(shortName="t", step=12, _hive_diag=diag)
     assert diag.file_count == 1
     assert diag.sel_count == 0
     assert len(r) == 6
@@ -217,7 +217,7 @@ def test_hive_sel_grib_1():
 
     # using hive partitioning keys + extra keys from GRIB header
     diag.reset()
-    r = ds.sel(
+    r = ds.to_fieldlist(
         {
             "shortName": "t",
             "step": 12,
@@ -234,9 +234,9 @@ def test_hive_sel_grib_2():
     root = earthkit_test_data_file("pattern/invalid")
     pattern = "_{shortName}_{date:date(%Y-%m-%dT%H:%M)}_{step}.grib"
 
-    ds = from_source("file-pattern", os.path.join(root, pattern), hive_partitioning=True).to_source()
+    ds = from_source("file-pattern", os.path.join(root, pattern), hive_partitioning=True)
 
-    r = ds.sel(shortName="t", step=12)
+    r = ds.to_fieldlist(shortName="t", step=12)
     assert len(r) == 0
